Rename file, export function, cleanup unused code
