--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "dotenv": "^16.3.2",
-    "mongoose": "^8.1.0",
-    "nodemailer": "^6.9.8"
-  }
-}
